feat: Add firebase_core and cloud_firestore dependencies
This commit adds the `firebase_core` and `cloud_firestore` dependencies to the `pubspec.yaml` file. The `firebase_core` dependency is updated to version 2.31.0, and the `cloud_firestore` dependency is updated to version 4.17.3. These dependencies are necessary for integrating Firebase services into the app.

The `build.gradle` files for Android and macOS are also updated to include the necessary configurations for Firebase. The `android/build.gradle` file adds the `google-services` classpath dependency, and the `macos/Flutter/GeneratedPluginRegistrant.swift` file registers the `cloud_firestore` and `firebase_core` plugins.

These changes enable the app to use Firebase services for features like real-time data synchronization and authentication.